requirements.txt
ebaysdk==2.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
python-dotenv==1.0.0
requests==2.27.0

